+++ odoo/local-src/connector_esb/models/backend_timestamp.py
@@ -33,6 +33,7 @@ class ESBBackendTimestamp(models.Model):
                     'promotion.alcyon',
                     'special.promotion',
                     'buyx.gety',
+                    'documents',
                     'stock.update']],
         string='Kind of export',
         readonly=True,