+++ HISTORY.rst
@@ -31,6 +31,7 @@ Unreleased
 **Bugfixes**
 
 * ALCN-1392: Delay only one export job when creating / modifying a sales order
+* Fix erroneous backend timestamp and kind
 
 **Build**